+++ src/app/(normal)/profile/page.tsx
@@ -1,13 +1,12 @@
 'use client';
 
-import ProfilePage from '@/app/(normal)/profile/[source]/[id]/page.js';
+import { ProfilePage } from '@/app/(normal)/profile/pages/Profile.js';
 import { NotLoginFallback } from '@/components/NotLoginFallback.js';
-import { SocialPlatform, SourceInURL } from '@/constants/enum.js';
 import { useCurrentProfile } from '@/hooks/useCurrentProfile.js';
 import { useUpdateCurrentVisitingProfile } from '@/hooks/useCurrentVisitingProfile.js';
 import { useGlobalState } from '@/store/useGlobalStore.js';
 
-export default function ProfileHome() {
+export default function Page() {
     const currentSource = useGlobalState.use.currentSource();
     const currentProfile = useCurrentProfile(currentSource);
 
@@ -17,12 +16,5 @@ export default function ProfileHome() {
         return <NotLoginFallback source={currentSource} />;
     }
 
-    return (
-        <ProfilePage
-            params={{
-                id: currentSource === SocialPlatform.Lens ? currentProfile.handle : currentProfile.profileId,
-                source: currentSource === SocialPlatform.Lens ? SourceInURL.Lens : SourceInURL.Farcaster,
-            }}
-        />
-    );
+    return <ProfilePage profile={currentProfile} />;
 }
